[Implementation] Blinking LED
Added definitions for pins and some other helper macros for the pins, namely pin number and bank.
Added a gpio struct to store the registers for general pupose I/O on the MCU.
Added a struct for the Reset and Clock Control unit to store the different registers. For blinking the LED, the AHB1 peripheral clock enable register is used. It turns GPIO banks on or off.
Added a function to write to a pin.
Added a function, spin(), to create a delay